Add Built In Custom Shaders to enable/disable scissor clipping
This adds a mechanism that we can use to define built in d3dx.ini
excerpts, which can include entire command list and custom shader
sections to allow us to ship a library of useful command lists & custom
shaders for whatever we see fit.

As the first pass, this is very simple - it just parses extra lines as
though they were part of the d3dx.ini and adds a "BuiltIn" namespace for
anything we wish to ship. It does not yet provide a means to ship the
actual shaders inside 3DMigoto, and these sections are currently logged
as though they were in the d3dx.ini.

This will be used to provide a "disable_scissor" option for backwards
compatibility that will call a simple command list that has the same
result as enabling/disabling scissor clipping on a per shader basis.

This may also be useful in the upcoming extension DLL support, to allow
an extension DLL to define new command lists & custom shaders on the
fly.