Make draw=from_caller work with hunting
In situations where we use the 'handling=skip' 'draw=from_caller' trick
to transfer a draw call from the game to a custom shader with a modified
state environment, we lose the ability to use the skip and mono marking
modes (the pink marking mode still works in this case), which although
it does make sense, is somewhat inconvenient and may be surprising to
users. Therefore, we want to skip the replacement draw call as well when
hunting.

We don't necessarily want to outright skip the entire command lists for
a selected shader in hunting mode, because we don't know what the
command list is doing, but we can probably fairly safely assume that
"draw = from_caller" should be skipped if the caller was skipped due to
hunting, so add a check to do so.