Add support for indirect and auto draw calls to "draw=from_caller"
If we are going to encourage more general use of draw=from_caller, it
will be wise to support all the possible draw calls that it may be used
with, so hook up the two indirect draw calls and DrawAuto.

This also makes the buffer used with the indirect draw calls available
via the "this" resource copy target.

This is not tested, as I am unaware of any games that use these calls.