Move the per-draw separation & convergence overrides to the command list
These have pretty limited use (realistically, they are only useful to
move something to either screen depth or infinity), and presets offers a
generally better way to set the convergence & separation, but they have
been used in the Defence Grid 2 fix at least so we can't just ditch
them.

Move them into the command list - this is partly for completeness, and
partly to remove some sequential dependencies in the BeforeDraw()
function so we can make marking_mode=mono work with draw=from_caller in
the next commit.

By default these are added to both pre and post command lists, where the
pre command list will save the separation/convergence, and the post
command list will restore it, and this will multiply by the old value
just as the old ShaderOverride code did (though, I question why we would
do that for convergence).

However, if an explicit command list is specified, this will set the
value and never restore it and will not perform the multiplication by
the old value - this gives us a little "value add" for moving it and
maybe someone can come up with a way to exploit this - an example that
comes to mind could use this to adjust the separation and convergence
during part of a frame to reduce the stereo effect on an off-screen
render target without making it fully mono.