Remove the old bit-rotted iteration feature from ShaderOverrides
The iteration feature has been broken forever - we never reset the
iterations on the draw calls, so it can only ever trigger once. It also
could only be used in conjunction with the per-draw call separation and
convergence overrides making it near useless and now that those have
been moved to the command list it will do nothing.

Remove it - in the future we will add support for incrementing ini
params in the command list to achieve something similar but a lot more
powerful.