Optimise BindStereoResources
Calling BindStereoResources from every draw call is costing a
significant amount of CPU time (2% in Grim Dawn) and reducing the fps in
CPU bound games.

Rather than continually rebind our resources, bind them once for each
HackerContext object, then from every SetShaderResources call check if
the game is attempting to unbind our resources and enforce that they
remain pinned. If this does happen we change the arguments so that we
can piggy back on the game's existing SetShaderResources call rather
than needing to call it again separately, and the result is that this is
reduced to more like 0.03% CPU.