Split Frame Analysis logging out of HackerContext
This splits the Frame Analysis logging (not dumping) code out of
HackerContext into its own context interface wrapper object. This new
object still lives within a HackerContext to keep the object management
(especially QueryInterface) simpler, and the HackerContext decides at
construction time whether to point mOrigContext to the
FrameAnalysisContext or not. The FrameAnalysisContext class implements
the ID3D11DeviceContext1 interface directly, so it can be used with
either a Context or a Context1 without having to maintain multiple
versions of the FrameAnalysisContext.

To keep the change "small" (relatively speaking), the HackerContext
still has a FrameAnalysisLog method, which just calls into the contained
FrameAnalysisContext.

We still allow frame analysis logging to be enabled dynamically at
runtime if hunting is changed from 0 to 1/2 by config reload, but this
will now only work for the immediate context - frame analysis logging on
deferred contexts will need a game restart to enable.

This change means that some of the frame analysis logging may change in
subtle ways - draw calls that we skip due to hunting or handling=skip
will no longer be logged, and calls where 3DMigoto has modified the
arguments will now show the modified arguments instead of the original
arguments from the game.

This change also means that some of the debug logging now happens a
little later, meaning we may miss the log for the current call if we
crash in HackerContext.

Swapping around HackerContext and FrameAnalysisContext could resolve
some of these issues, but would result in the calls in the frame
analysis log not being in a logical order (e.g. we want to log the pre
command list, then the draw call, then the post command list because
that is the logical order they run in, but swapping these contexts
around would mean that we would have to log the draw command before the
pre command list, or after the post command list).

HackerContext now contains three ID3D11DeviceContext objects to choose
from (not counting the Context1 versions):

- mOrigContext should be used for the *original* DirectX calls that we
  want logged through Frame Analysis and/or the debug log.

- mRealOrigContext will point to the actual DirectX object we are
  wrapping, which may or may not be hooked (and if it is hooked it will
  call back into HackerContext). This is the object generally expected
  by code outside of the context when it is the object that is of
  interest, as opposed to making DirectX calls.

- mPassThroughContext allows calls into DirectX that will always bypass
  HackerContext and frame analysis logging, which may either be the
  original DirectX object, or the trampoline object if hooking is
  enabled.

This results in about a 0.5% reduction in CPU usage, which improves the
framerate in Grim Dawn by about 1fps - a fairly invasive change for a
relatively small improvement, but a measurable improvement nonetheless.