Overlay & Hunting: Use PassThrough context
The overlay and screenshot on mark was using the original DirectX
context object to bypass 3DMigoto's HackerContext, however when hooking
is enabled this object will call back into HackerContext so 3DMigoto
will intercept itself.

Change it to use the pass through context, which will use the trampoline
object if hooking is enabled to bypass 3DMigoto.