Fix crash in TheDivision on Win10.
This is a game and OS specific crash, where it somehow knows that it is on Win10, and requests a QueryInterface(IDXGISwapChain2), which does not exist in our SDK 8.0 code base.

We return the 'this' object, but that is not sufficient for it, even though the pointer to our mOrigUnknown and the result of the QueryInterface are identical.

It jumps off to a random call, most likely because the vtable does not match what they expect.

On Win10, even though the object pointers are identical, it does not appear to be true that the vtables match.

This is a one-off fix for TheDivision, not sure this is the best approach, but it succeeds in making them not call whatever jumps off into funny space.