First part of restructing hooking in DLLMainHook.
This compiles, but causes other changes for refactoring in other files.

The DLLMainHook environment was contaminated with a lot of extra code that doesn't belong in this unit, because the runtime environment is completely different than normal code.

Anything that can be installed after the fact, like features to be enabled, needs to be in different compile units.

Even though we use in-proc hooks for other purposes, that is not the important part of  this file.  The important part for this file is that it is DLLMain time, when the d3d11.dll has been loaded, but not executed.  This is a restricted runtime and should only include things that are strictly required at launch.

The actual hooked code has also been moved to their respective compile units, because they confuse the idea of the strict runtime.  So, Hooked_LoadLibraryExW and Hooked_DXGIFactory have been moved.  Nothing should ever call *into* the DLLMainHook.

It is important to keep the DLLMainHook strict and simple, because mistakes here can cause games to crash or hang at launch.