Adding dxgi.lib to all build targets.
The reason to add this as a fixed dependency is to force the dxgi.dll to be loaded as early as possible, in DLLMainHook.

We need that to load as early as possible so that we can hook CreateDXGIFactory and Factory1.  Otherwise a game can call those before we get a chance to hook.

No game can run without dxgi.dll, so adding this dependency should not cause any problems.