Force dxgi.dll to be loaded at game launch.
This uses the "/INCLUDE IDXGICreateFactory" in the project settings to specify that symbol to be included.

Without this tweak, even though we reference IDXGICreateFactory for hooking, we never call anything directly in dxgi.dll and it will not be hard linked to the d3d11.dll.

We want it to be hard-linked so that it is loaded at launch.  That allows us to capture any early calls to IDXGICreateFactory that might happen before any d3d11.dll calls.