Move MouseHooks and WindowPosHook.
MouseHooks are moved to where they are used, in IniHandler.

WindowPosHook is moved to where it is used, in HackerDXGI.

These are both standard runtime, and may or may not actually be hooked, depending upon the settings in d3dx.ini.

Functionally, this should be identical to how it was.

Compiles and runs, although I'm not sure how to test these functions.