Create the Present hook for SwapChains.
This creates a temporary Window, Device, and SwapChain, in order to get access to the Present call address.

Using that address, it installs a Hooked_Present call, which is now called for every game Present.

This is tested as functional in the game Inside.  Hook is installed and called without errors.