Successfully implement hooked-only Present.
Moves all the RunFrameActions functionality into HookedDXGI.cpp, and removes any dependency upon the HackerDXGISwapChain.   HackerDXGI.cpp is removed from project.

Makes 4 new globals for references to the HackerDevice, HackerContext, IDXGISwapChain, and Overlay.  These are used by the Present hook to draw the Overlay and passed along to the unchanged RunFrameActions.

A couple of smallish changes to use the G->gSwapChain directly, the original game version, to do screenshots.

Barely tested.  This does function correctly in Inside, overlay is drawn, hunting works. Proof of concept.

This does break a couple of features- Upscaling is non-trival to put into this framework and requires more hooks. ForceDisplayParams for fullscreen only works if it comes off the CreateDeviceAndSwapChain, and needs new CreateSwapChain, CreateSwapChain1, CreateSwapChainHwnd hooks.