Switch to daisy-chain model for installing hooks.
The prior variant would directly create a SwapChain to be able to get access to the Present function address.  That required building a Device.

On Win7, that works.  On Win10, that CreateDevice calls back to the CreateFactory, which is hooked and tries to create another Present hook, and recurses until stack dies.

This approach is to install hooks in a daisy-chain fashion. Factory is done at DLLMainHook.  When called, it creates a hook for CreateSwapChain.  When CreateSwapChain is called, it creates a hook for Present.

This avoids the possibility of getting called back, and also makes CreateSwapChain available, which allows us to revive the ForceFullScreen functionality.

Tested on Inside on Win10 and Overlay works as well as hunting.