Add _CreateDXGIFactory@8 reference to x32 projects.
For whatever freaky Microsoft reason, the dxgi.lib does not match from x32 to x64.  In x32 it has this badly formed label instead.

In order to force load our dxgi reference, we still need something in dxgi to refer to, and this is it for x32.

In x64 it is simply CreateDXGIFactory.