Remove all HackerDXGI references.
Since we are just doing hooking now, all references to HackerDXGI objects are obsolete.

This includes the skip_dxgi_factory and skip_dxgi_device functionality, because with hooking we do not need to skip.

This removes the 'secret' path to the IDXGIFactory through HackerDevice->QueryInterface(IDXGIDevice) where it daisy chains over to IDXGIFactory through QueryInterface calls.  No longer necessary to catch these QueryInterface, because the CreateSwapChain is always hooked.

Cleanup some comments.