Fix crash in overlay code if shader info line is too long
Similar to the previous crash we hit in the ini parser with wcscpy_s,
using wcscat_s can also crash us if the destination buffer is too small,
which can happen in the overlay code if a shader info line is too long.

Similar to the wcscpy_s crash, the fix is to switch to wcsncat_s and
pass the _TRUNCATE option, telling it that it is OK to truncate the
resulting string to fit in the destination, since we won't be able to
fit a long string on screen anyway.