Command Lists: Introduce mechanism to set convergence from GPU
This introduces a ResourceStagingOperation capable of asynchronously
transferring a resource from the GPU to the CPU, and enables the
separation & convergence commands to use it to set the stereo settings
from values in a buffer on the GPU, that could be filled out by a shader
implementing some form of auto-convergence algorithm, e.g. by finding
the minimal value on the depth buffer.

The value is transferred asynchronously, meaning that the separation /
convergence will not be set every time it is called - rather, the first
time it is called it will initiate a transfer from the GPU, and
subsequent calls will only set the separation/convergence if the value
has finished transferring to the CPU. Once the value is available, the
following call will start a new transfer. Only one transfer can be
active at a time - there is currently no provision to have multiple
values in flight at a time (and we would have to ensure we have a sane
limit if we implement this in future, for cases where the framerate is
unbounded and insanely high).

e.g.

    [ResourceAutoConvergence]
    type = Buffer
    format = R32_FLOAT
    array = 1

    [CustomShaderAutoConvergence]
    cs = ShaderFixes/auto_convergence.hlsl
    cs-u0 = ResourceAutoConvergence
    post cs-u0 = null
    dispatch = 1, 1, 1

    [Present]
    pre run = CustomShaderAutoConvergence
    pre convergence = ResourceAutoConvergence

Note that specifying "pre" (or post) is necessary to persist the value
when setting separation or convergence, as omitting this would restore
the old value after the command list it is set. This is for
compatibility with the way 3DMigoto historically handled the separation
and convergence settings in ShaderOverride sections.