Refactoring HackerContext and HackerDevice for *1.
This completes the two big files to use the ID3D11Device1 and
ID3D11DeviceContext1 as their primary composition object.

Fixes for QueryInterface to return 'this' for either Device1 or Device. Or Context1 or Context.  We are always going to return
the highest level object we expect to make.

Changes all the mOrigDevice references to mOrigDevice1 references, to keep object type more clear.  Similarly for all the other composition objects.

For GetImmediateContext or CreateDeferredContext, we cast to our *1 objects, and create a wrapper HackerContext.

Compiles, but is untested.