Update CreateDevice to always return Device1.
Similarly for CreateDeviceAndSwapChain.  This takes out the
platform_update check, and just assumes creation of the
higher level object always.

It will upcast from ID3D11Device and ID3D11DeviceContext to ID3D11Device1 and ID3D11DeviceContext1 using QueryInterface, always.  If we get an error we still store the result as *1, which should only happen on Win7 with no platform update installed.

Compiles, but is untested.