Replace all mOrigDevice with mOrigDevice1.
And similarly for all the other mOrig* composition objects.

This seems worthwhile to keep clarity on what precisely a given reference is too.  It seems OK to keep HackerDevice without a 1, because it is clear that it is our single layer object wrapping a Device1.

With this, there is only one time where we lie, which is for the Win7 without platform update installed.  We will still store the normal Device in our mOrigDevice1 composition object.  We might need to revisit this if it causes problems.