Auto convergence: Begin next transfer as soon as the previous one completes
After a scene change, the auto-convergence feature requires at least one
frame drawn with the old convergence value for it to determine the new
value to use. This is unavoidable if using either the depth buffer or
stream output from the previous frame to set the convergence for the
next frame, but we ideally want as few frames displayed with wrong
convergence as possible, so begin a new transfer as soon as we have
retried the previous one to minimise this.