Disable hooking temporarily.
Temporarily disable hooking in order to get it to compile.
Even more refactoring needs to be done with hooks.

Using this variant, compiles and runs.

Tested in ShadowOfMordor and it seems to run correctly
and fix the game.  Tested in Just Cause 3, and it crashes,
but that game is super flaky.  Might be unrelated.

Tested in BatmanTelltale, and it actually does not
crash.  DXGI hook is disabled because they are looking
for CreateSwapChainForHwnd, but the part we are
interested in here- the single layer HackerDevice and
HackerContext are both working.