Add Present1, CreateSwapChainForHwnd, Factory2 hooks.
For BatmanTelltale, it calls through the secret path to fetch
an IDXGIFactory2 via the Device parent.  This sequence
upcasts any call to Factory to Factory2, and installs all Create
hooks for:
QueryInterface, CreateSwapChain, CreateSwapChainForHwnd.

QueryInterface is also hooked, because they might take a Factory1 and upcast it to Factory2.  Not positive this is necessary now because the original CreateFactory will upcast and hook that scenario.

In any case, this successfully creates an Overlay for the BatmanTelltale game.  Game crashes still, but DXGI is clearly hooked.