Command List: Add ability to set ini params to stereo params
This allows an ini param value to be set to one of the stereo
parameters. This may seem a little redundant at first given that is
pretty close to what StereoParams is for, but there is a small
difference that StereoParams is only updated once per frame, where this
will get the current values from nvapi, regardless of how they may have
changed since the start of the frame.

This is mostly intended for use after staging the convergence from the
GPU, where we don't otherwise know if the operation has completed or
not, and when we want the value prior to the next present call.

For example, an auto-convergence algorithm may want to be able to take
user convergence adjustments into account, and to do so it will need to
compare snapshots of the convergence from before and after the present
call without having to be concerned with how it itself may or may not
have affected the convergence during the frame. Since the convergence is
updated asynchronously, it could not simply remember the previous
convergence value it had set, since it doesn't know which (if any) frame
that would be applied to.