Custom Resources: Add data= to specify initial data for buffers
This allows the initial data to be specified when creating a custom
buffer. In order for this to work, a supported format must also be
specified for the buffer, as that format will determine how the data
field is parsed and how the initial data field is filled out. The format
may either be specified with the format= setting, or as the first
parameter in the data= line, e.g.

[ResourceBuffer]
type = Buffer
format = R32_FLOAT
data = 0 1 2 3 4 5 6 7 8

[ResourceStructuredBuffer]
type = StructuredBuffer
data = R32_FLOAT 0 1 2 3 4 5 6 7 8

This supports a wide range of simple format types. All fields have to be
the same size (e.g. R32G32B32A32 is OK, but D32S8X24 is not) and no
field can be TYPELESS or unused (no X fields). 32-bit floating point
formats are supported, but 16-bit and other unusual float sizes are not.
1, 2 and 4 byte signed and unsigned integer formats are all supported,
and 1 and 2 byte normalized integer formats (SNORM and UNORM) are
supported. Sub-byte formats are not supported, nor are block compressed
or other esoteric formats.

The initial data for complex structured buffers cannot be specified
directly yet, but in many cases we can still initialise these by
pretending they have a simpler format. This fake format must be
specified as the first parameter in the data= line, not in the format=
line (using the later may prevent the buffer from being bound as a
view, as the view would attempt to use that format). A given data entry
can be specified as a hex string, which can be used to embed a different
type to fake a more complicated data structure.

This provides another alternative to manually specifying the size of a
buffer, which will be automatically determined from the format and
initial data. Structured buffers will also have their stride set to the
size of the initial data by default. These still be manually specified
to override this behaviour, and if the initial data has to be enlarged
(or created) the uninitialised portions will now be zeroed out.