Fix theoretical refcounting bug when saving off class linkage
When we save off the class linkage, we don't bump its refcount, and the
game could then free the original object. If we later reuse the linkage
when reloading shaders or patching a shader with ShaderRegex we would
potentially pass an invalid pointer to the compiler, leading to a crash.

Bump the refcount when saving the class linkage to fix this, though as
I've never actually seen a game use class linkage this bug is very
theoretical and this fix is very untested.