Merge different types of shader maps together
We don't really need separate maps to store the original shaders or
hashes for each different type of shader, because the memory handles we
use as keys for these should* be unique independent of the shader type,
and since they are stored in unordered maps the lookup time will be
constant regardless of how many entries are in each map.

Merging these reduces our code complexity a little and gets rid of a lot
of template arguments, which is a good thing. We could still probably go
a little further and merge more of the remaining maps together, but that
is a little more of an invasive change so leaving it for now.

* Barring known handle reuse bugs, which will be fixed in a following
commit - merging these maps now will simplify that fix later.