Fix shaders occasionally being randomly mixed up on F10 reload
Occasionally a random shader may appear to break after pressing F10 to
reload shaders. This can manifest in a variety of different appearances,
from certain objects or the entire screen turning black or going
nuclear, flickering, disappearing, etc. Essentially, 3DMigoto replaces
one shader with a totally unrelated shader causing all sorts of
weirdness.

This is caused by a sequence similar to the following:
1. The game creates a shader which 3DMigoto replaces out of ShaderFixes
2. The game releases the shader
3. The game creates another shader of the same type that 3DMigoto does
   *not* replace from ShaderFixes, that by chance ends up at the same
   address as the shader in step 1
4. The user presses F10 to reload shaders
5. 3DMigoto mistakenly thinks that the shader created in step 3 was
   loaded from ShaderFixes, but since it is not able to find it there it
   attempts to revert it, but uses the original shader from step 1
   instead

This problem is due to the various data structures 3DMigoto holds about
shaders from the game holding stale data from previous shaders that may
have previously occupied the same memory address. Fundamentally, this
occurs because we do not wrap or hook the shader's Release() call, and
cannot clean up our data structures. In most cases this is OK, because
when the game creates the new shader we update our data structures
overwriting most of the stale data, but in this situation we missed
wiping out stale data from one of the maps as we didn't think we needed
it for the new shader.

This takes a pretty big hammer to fix this - whenever we create a
shader (for any reason), we check if the handle already exists in any of
our shader maps and erases it if it is, freeing any associated objects
we had saved with it. We do it at this point, because we can be certain
that the handle returned by CreateShader was invalid prior to the call,
so if that handle exists in any of our maps it was definitely stale.

An alternative less heavy-handed approach would be to just wipe out the
stale entries when the game creates a shader, but I'm concerned that
might miss some rarer edge cases, e.g. suppose an application called
XXGetShader() and retrieved a shader that 3DMigoto had set (as in, one
of our shaders, not one of the game's shaders) while saving some state,
then later called XXSetShader() to restore that state - we would look up
that handle, and if that handle happened to have been previously used by
a shader we replaced from the game we would go ahead and replace it
again. That would be exceedingly rare, but theoretically possible -
cleaning up our maps after any CreateShader() call should avoid this.

This commit fixes this bug:
  https://github.com/bo3b/3Dmigoto/issues/86