Remove Hacker* requirements from Overlay.
The Overlay just needs to draw, and we were saving
references to HackerDevice and HackerContext.

With current changes, the constructor is getting just
the original Device and Context, so as to remove the
need to dereference those to get the original objects.

The only part we actually need is a HackerDevice->mStereoParams
reference, which we can now fetch from the global.