Add ForceDisplayParams1 for CreateSwapChainForHwnd.
The CreateSwapChainForHwnd uses yet another different
swap chain descriptions, including a special purpose
full-screen variant, so prior code does not work.

This is adding the routine to handle forcing a game to
full screen in a platform_update scenario.  This is required
for Batman Telltale at least.

The other forcing functions are skipped for now, as the
input formats are completely different and it's not obvious
how to properly force other things like refresh rate and
even screen size in the CreateSwapChainForHwnd case.

Stub routines for those features will log an error and BeepFailure for notice.  Will require experimentation.