Create Overlay for CreateSwapChainForHwnd.
In this case, there can be no direct path from CreateDeviceAndSwapChain, because IDXGISwapChain1 types are only ever created from Factory2.

This adds the similar code to CreateSwapChain where it dereferences the input HackerDevice, and creates the Overlay with all Orig objects.

Works in BatmanTT on Win10, one game where we expect to see this.