Assembler: Fix blank lines with DOS style newlines corrupting the shaders
Helifax ran into an issue where a blank line before the shader model
could silently corrupt corrupt the shader if the assembly was using DOS
style newlines:

  https://forums.geforce.com/default/topic/1029242/3d-vision/mass-effect-andromeda-3d-vision-fix-placeholder-/post/5274222/#5274222

Examination of the resulting binaries showed that the shader model had
been changed from cs_5_0 (0x50000500) to ps_0_0 (0x00000000).

The assembler is supposed to pre-strip carriage returns before
processing lines, but an off by one error meant that it would not strip
carriage returns if they were the only character on the line. This would
mean that it prematurely believed the code had started while looking for
the shader model, and would attempt to assemble the blank line as the
shader model, which would silently fail and not insert any instruction.
The assembler would then insert a 0x00000000 as a placeholder for the
shader size, now taking the place of the shader model. The real shader
model would then be assembled and placed where the shader size was
supposed to go, and would later be overwritten by the real shader size.

This fixes the off by one error so that carriage returns will be
stripped from blank lines as well.

Reported-by: Helifax