Revert "Split Frame Analysis logging out of HackerContext"
It was expected that splitting the frame analysis logging out of
HackerContext may degrade the quality of the frame analysis log, however
this turned out to be a little too severe to justify the minor ~1fps
performance benefit of doing so.

The most severe issue turned out to be that XXSetShader calls would be
missing the shader hash if the shader had been reloaded, making the log
all but useless to trace resources from such a shader - and given that
was it's primary purpose, this is not an acceptable compromise.

We can look at bringing this back later - this issue could be
alternatively addressed by promoting FrameAnalysisContext to run before
HackerContext. Doing so will change the order of the command list
logging with respect to the log of the original call, but we should be
able to tweak the logging to cope with that. Making that change at the
moment would be difficult because of the complex interaction between the
game, FrameAnalysisContext, HackerContext, HookedContext and the
Context1 variants of all three. Once we have moved to the single layer
model we can revisit this, as it will have eliminated a large amount of
the complexity.

This reverts commit 1c3fcab97cdfefb839c5e6b1af0bf2606e6f3119.