Remove mPassThroughContext1
The new approach to FrameAnalysisContext means that we no longer need to
distinguish between mPassThroughContext1 and mOrigContext1 inside the
HackerContext - this was necessary when HackerContext could optionally
call into FrameAnalysisContext to keep track of whether it was calling
the original context or FrameAnalysisContext, but since HackerContext no
longer knows about FrameAnalysisContext it no longer needs to track
this. Remove mPassThroughContext1 and change any uses of it to use
mOrigContext1 instead (this is effectively what we used to do before
FrameAnalysisContext was split out in the first place, so this matches
the reversion in master as well).