Frame Analysis: Log Dispatch and SetShader calls before HackerContext
This moves the log for Dispatch, DispatchIndirect and all the SetShader
calls to before calling into HackerContext.

The Dispatch logs were originally made between the BeforeDispatch and
AfterDispatch calls, which is not possible now that FrameAnalysisContext
has been split out, and moving them before calling into HackerContext
makes them consistent with the Draw calls - of particular importance is
that this corrects the draw call number in these logs.

The SetShader calls originally had logging done afterwards as they
logged the shader hash, and needed a template function to have completed
to update the mCurrentXXXShader members. With the split out
FrameAnalysisContext they no longer use this and the logging function
gets the hash itself, so it is no longer necessary to log them after the
original call and better to log them first in case we get a crash inside
HackerContext or the original call so we see them in the debug log.

The Get functions are still all logged after calling into HackerContext
so that they can log the result. It might be a good idea to consider
moving them to log before HackerContext as well, but that will need a
little more consideration as to how to log the result, so that can come
later.