Command List: Tweak logging to show recursion & pre/post
In 1.3 we will be splitting out FrameAnalysisContext from HackerContext
once again and running it first, which means that the command list log
will be out of order with respect to the call that triggered it.
Currently the pre command list would be logged before the original call,
and the post command list logged after it, but in 1.3 they will both be
logged after it, and there won't be any other indication in the log to
show where the pre command list finished and the post command list
started to infer this either.

We can't easily resolve the order, but we can tweak the logging to show
the pre and post command lists, from which it can be inferred that the
original call would go between the two - this adds additional logging
when starting to run a command list to show this in preparation for 1.3.

While we're at it, since command lists can recursively call each other,
let's add variable indentation to show the current level of recursion.
This is useful to see regardless of the changes in 1.3, so it can go in
1.2 as well.