Delete all Hacker* objects except HackerSwapChain.
Delete all other DXGI object wrappers, except for the primary
one of IDXGISwapChain1.  All others cannot be wrapped at
a lower level, we can only use a single-layer wrap, so all
code needs to be in the HackerSwapChain.

HackerDXGISwapChain is renamed to HackerSwapChain, because really, what other kind than DXGI is there?

This does not compile, but is the broad brush stroke of the refactoring.