Move all Present logic back to HackerDXGI.
Since we are going to have a single-layer IDXGISwapChain1
wrapper, we need the logic back in HackerSwapChain::Present

Also deletes the hook for Present and Present1 from
HookedDXGI, because the wrapper will receive these calls.

Removes use of globals from RunFrameActions and
UpdateStereoParams, because they are now methods
of HackerSwapChain.

Does not compile.