Return HackerSwapChain object, and set global.
In CreateSwapChain hook, we now wrap that object with a
new HackerSwapChain, and return that to the game. This
will allow our HackerSwapChain::Present call to be called.

Also sets up the global for gHackerSwapChain for use in
driving our UI.

Does not compile, headers are all botched up with circular
loops.