Fill out all fields of mResources all the time
We were skipping filling out some fields of mResources for Buffers and
Texture1Ds, but this was done to prevent hash contamination detection
and tracking on these resources, and is not really the right place to
disable that, especially since it would also prevent us from using stat
collection on these resources.

Fill out everything relevant to the resources, and add a resource type
field that the hash contamination & tracking code can test early.