Fix crash on ResizeBuffers() if overlay has ever been displayed
The overlay did not release it's view of the back buffer, which can
cause ResizeBuffers() to crash in some games if the overlay had ever
been displayed since launch. This includes Mass Effect Andromeda, but we
have also seen this same crash in many other games if the command list
cached a view of the back buffer (which we now forbid), and this may
explain other crashes we have seen in some games from time to time.