ShaderUsage: Remove <PeerShaders> from <ComputeShaders>
Compute shaders don't have peer shaders, so don't add an entry for them.
Rather than passing another flag to indicate which shaders have peers,
just check if the PeerShaders set has anything in it. In fact - get rid
of the other flags as well, since if the data structures are empty
nothing will be printed and this simplifies the code.