ShaderUsage: Fix showing incorrect hashes if a resource handle is reused
The stat collection saved off the memory address of any resource it was
going to dump, but since we do not keep a reference to that resource it
could be released by the game, and the handle reused. If this happens,
then when we dump out the ShaderUsage.txt we will look up the mResources
data structure to find the resource hash, but we may get an unrelated
one.

Save the hashes along with the resource handles to make sure that they
are correct in the ShaderUsage.txt.