Switch static_cast to reinterpret_cast.
reinterpret_cast is more clear, and specifically says we are reinterpreting the object as something else.  static_cast is generic and doesn't add anything.  Both should have the same effect, and no performance impact.

Switching to this form because multiple examples from Chuck Walbourn use this style, and that's compelling enough for me.