ShaderUsage: Use naming consistent with syntax in d3dx.ini
The ShaderUsage.txt is the main place we expect users to grab the
resource descriptions to match in TextureOverride sections, so to
simplify things change them to use the same naming scheme.

This also changes usage to output a text string, since the match_usage
doesn't accept numeric inputs. It would be nice to do the same to the
flags fields as well, but that's a little more involved and those fields
will accept hex strings, so leaving that for later.