Remove Path1 from Hooked_CreateSwapChain.
Path1 was from the CreateDeviceAndSwapChain implicit hook.
That path has been broken into two pieces, and not calling
through to the original, so now we only need the primary
path through CreateSwapChain, which will now always
receive a HackerDevice on input.

Overlay creation has been moved back to the constructor
for HackerSwapChain.

This needs further cleanup, and is a testing state.
But, tested in Witcher3 without any errors, and Overlay
works.