Remove use of Overlay->Resize by updating each frame.
Dishonored2 does an unusual thing by setting the framesize
to 0,0 and not restoring it.  We would set Overlay to 0,0 and
thus not draw.

There is no real need to keep the resolution for the Overlay,
because we always have to fetch the backbuffer in order
to be able to setup to draw.  Since we have the backbuffer,
this change just fetches the current size at each InitDrawState.

Resize is therefore no longer necessary, it will always
draw to the correct size, and always be in sync with the game.

Tested to work in Dishonored2 and BatmanTT.